refactor(web): extract service tier for taxonomy.ts (closes #3329)
Follow-up to #3231 / PR #3335. Extracts the implementation of
`suggestOccupations`, `suggestSeniorities`, `suggestTechnologies`,
`resolveOccupationSlugs`, `resolveSenioritySlugs`,
`resolveTechnologySlugs`, `expandOccupationIds`,
`expandOccupationIdsBatch`, `getAllOccupationsGrouped`,
`getAllSeniorities`, `getAllTechnologiesGrouped`,
`getEmploymentTypeCounts`, and `getWorkModeCounts` from
`apps/web/src/lib/actions/taxonomy.ts` into
`apps/web/src/lib/services/taxonomy.ts` (new — server-only module,
no `"use server"` directive).

`apps/web/src/lib/actions/taxonomy.ts` becomes a thin `"use server"`
wrapper of DECLARED async functions that delegate to the service tier
so existing UI callers (search-bar typeaheads, modal counts, watchlist
resolve flows) keep working with zero churn. Pure `export { foo } from
"..."` would break the Turbopack production build at every client
component that imports the wrapper (PR #3335 regression) — so each
callable is an explicit `export async function foo(...)` delegation
instead.

Route handlers `/api/v1/taxonomies` and `/api/v1/resolve` now import
directly from `@/lib/services/taxonomy` so they don't pay the
server-action machinery cost (per-call RPC URL, serialization
boundary, security IDs) for what is already a public REST surface.

`apps/web/src/lib/services/search-input.ts` previously reached back
into `@/lib/actions/taxonomy` from inside the service tier — closed
that boundary leak.

Test mocks at `@/lib/actions/taxonomy` were swapped for
`@/lib/services/taxonomy` only in the two tests where the production
import path changed (`search-input.test.ts` exercises
`services/search-input` -> `services/taxonomy`; resolve route test
exercises `route.ts` -> `services/taxonomy`). All other consumers
(modals, watchlists, company) still import from the action wrapper
and their mocks continue to intercept that path.

New boundary test at `services/__tests__/taxonomy-tier.test.ts`
mirrors `search-tier.test.ts`:

- service file has `import "server-only"` and no `"use server"`
- action wrapper keeps `"use server"`
- service and wrapper expose the same callable surface
- wrapper declares an explicit `export async function ${name}` and
  `await/return service.${name}(...)` delegation for every service
  callable
- wrapper contains no `export { foo } from "@/lib/services/..."`
  value re-exports

Gates: `pnpm exec tsc --noEmit` clean, `pnpm test` 1195 pass (1190
pre-existing + 5 new in `taxonomy-tier.test.ts`),
`pnpm --filter @jobseek/web run build` clean.
