+++ apps/web/app/api/v1/resolve/route.ts
@@ -1,10 +1,15 @@
 import { type NextRequest, NextResponse } from "next/server";
+// Public REST routes import the plain service tier (`@/lib/services/*`)
+// rather than the `"use server"` action modules (`@/lib/actions/*`). The
+// service functions are functionally identical but avoid the
+// server-action machinery (per-call RPC URL, serialization boundary,
+// security IDs). See issues #3231 / #3329.
 import { suggestLocations } from "@/lib/actions/locations";
 import {
   suggestOccupations,
   suggestSeniorities,
   suggestTechnologies,
-} from "@/lib/actions/taxonomy";
+} from "@/lib/services/taxonomy";
 import { suggestIndustries } from "@/lib/actions/company";
 import { CACHE_TTL_LONG } from "@/lib/cache-ttl";
 import { slugifyTitle } from "@/lib/watchlist-slug";
